Enable the update of product quantities in the cart
We created an instance of CartAddProductForm for each item in the cart to allow changing product quantities. We initialize the form with the current item quantity and set the override field to True so that when we submit the form to the cart_ add view, the current quantity is replaced with the new one.

We added {{ item.update_quantity_form.quantity }} {{ item.update_quantity_form.override }} to our detail view so that they can be part of the form being submitted for the update since they are already form elements defined in forms.py