Create a context processor and add it to the settings
A context processor is a Python function that takes the request object as an argument and returns a dictionary that gets added to the request context. Context processors come in handy when you need to make something available globally to all templates.

Context processors are executed in all the requests that use RequestContext. We might want to create a custom template tag instead of a context processor if the functionality is not needed in all templates, especially if it involves database queries.

https://docs.djangoproject.com/en/3.2/ref/templates/api/#built-in-template-context-processors

https://docs.djangoproject.com/en/3.2/ref/templates/api/#django.template.RequestContext